refactor(modeling - menugroup) : 메뉴그룹 패키지 구조 변경
